gcloud/upload_to_storage/requirements.txt
google-cloud-storage==1.7.0
google-cloud-logging==1.4.0
google-cloud-error_reporting==0.29.0
oauth2client==4.1.2
six==1.11.0
requests[security]==2.18.4
